+++ apps/meteor/app/livechat/server/hooks/afterUserActions.ts
@@ -2,7 +2,6 @@ import { type IUser } from '@rocket.chat/core-typings';
 import { Users } from '@rocket.chat/models';
 
 import { callbacks } from '../../../../lib/callbacks';
-import { Livechat } from '../lib/Livechat';
 import { Livechat as LivechatTyped } from '../lib/LivechatTyped';
 
 type IAfterSaveUserProps = {
@@ -21,14 +20,14 @@ const handleAgentUpdated = async (userData: IAfterSaveUserProps) => {
 	}
 
 	if (!wasAgent(oldUser) && isAgent(newUser)) {
-		await Livechat.afterAgentAdded(newUser);
+		await LivechatTyped.afterAgentAdded(newUser);
 	}
 };
 
 const handleAgentCreated = async (user: IUser) => {
 	// created === no prev roles :)
 	if (isAgent(user)) {
-		await Livechat.afterAgentAdded(user);
+		await LivechatTyped.afterAgentAdded(user);
 	}
 };
 
@@ -39,8 +38,8 @@ const handleDeactivateUser = async (user: IUser) => {
 };
 
 const handleActivateUser = async (user: IUser) => {
-	if (isAgent(user)) {
-		await Livechat.addAgent(user.username);
+	if (isAgent(user) && user.username) {
+		await LivechatTyped.addAgent(user.username);
 	}
 };