+++ apps/meteor/client/startup/actionButtons/index.ts
@@ -5,7 +5,6 @@ import './jumpToStarMessage';
 import './permalinkPinned';
 import './permalinkStar';
 import './pinMessage';
-import './readReceipt';
 import './starMessage';
 import './unpinMessage';
 import './unstarMessage';