+++ apps/meteor/client/views/root/AppLayout.tsx
@@ -5,7 +5,6 @@ import { useAnalytics } from '../../../app/analytics/client/loadScript';
 import { useAnalyticsEventTracking } from '../../hooks/useAnalyticsEventTracking';
 import { appLayout } from '../../lib/appLayout';
 import PageLoading from './PageLoading';
-import { useCloudAnnouncements } from './hooks/useCloudAnnouncements';
 import { useEscapeKeyStroke } from './hooks/useEscapeKeyStroke';
 import { useGoogleTagManager } from './hooks/useGoogleTagManager';
 import { useMessageLinkClicks } from './hooks/useMessageLinkClicks';
@@ -24,7 +23,6 @@ const AppLayout = () => {
 	useAnalytics();
 	useEscapeKeyStroke();
 	useAnalyticsEventTracking();
-	useCloudAnnouncements();
 
 	const layout = useSyncExternalStore(appLayout.subscribe, appLayout.getSnapshot);