+++ apps/meteor/tests/e2e/page-objects/omnichannel-manager.ts
@@ -12,6 +12,20 @@ export class OmnichannelManager {
 		this.sidenav = new OmnichannelSidenav(page);
 	}
 
+	get inputSearch() {
+		return this.page.locator('[placeholder="Search"]');
+	}
+
+	async search(text: string) {
+		await this.inputSearch.fill(text);
+		await this.page.waitForTimeout(500);
+	}
+
+	async clearSearch() {
+		await this.inputSearch.fill('');
+		await this.page.waitForTimeout(500);
+	}
+
 	get inputUsername(): Locator {
 		return this.page.locator('input').first();
 	}