+++ packages/models/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @rocket.chat/models
 
+## 0.0.21-rc.7
+
+### Patch Changes
+
+- @rocket.chat/model-typings@0.2.0-rc.7
+
+## 0.0.21-rc.6
+
+### Patch Changes
+
+- @rocket.chat/model-typings@0.2.0-rc.6
+
+## 0.0.21-rc.5
+
+### Patch Changes
+
+- @rocket.chat/model-typings@0.2.0-rc.5
+
+## 0.0.21-rc.4
+
+### Patch Changes
+
+- @rocket.chat/model-typings@0.2.0-rc.4
+
+## 0.0.21-rc.3
+
+### Patch Changes
+
+- @rocket.chat/model-typings@0.2.0-rc.3
+
 ## 0.0.21-rc.2
 
 ### Patch Changes
@@ -21,6 +51,24 @@
 - Updated dependencies [ec1b2b9846]
 - Updated dependencies [5f81a0f3cb]
   - @rocket.chat/model-typings@0.2.0-rc.0
+
+## 0.0.23
+
+### Patch Changes
+
+- @rocket.chat/model-typings@0.1.8
+
+## 0.0.22
+
+### Patch Changes
+
+- @rocket.chat/model-typings@0.1.7
+
+## 0.0.21
+
+### Patch Changes
+
+- @rocket.chat/model-typings@0.1.6
 
 ## 0.0.20